Analytic event is sent to GameAnalytics server.
Events appear on the "Live Events" panel on the web client. Problem was there was a missing DLL in the addon folder. May want to try deleting the remaining files later, as I just put the entire repository into our project under the "game_analytics" folder.
Next: Figure out how to organize events on the dashboard to figure out what's going on with our game.