Slightly improved Firework Flora performance.
There is less of a lag spike each time the powerup is shot. There is something weird with the way caching is done. The solution to eliminate lag entirely would be to use object pooling, but this is a big change that I don't want to do right now. Hopefully it is not needed, but it probably will be if we want the game playable on low-end computers.