Slightly simplified Enemy._physics_process logic.
Trying to fix lag. So far, I know that move_and_slide() causes most of the lag with Enemies. This is especially the case when a lot of them are grouped up and collide with each other. For continuous damage powerups, part of the lag comes from having to instantiate so many damage effects.